Initialize Repository
This game was developed through the unity tutorial which provided the
assets and explanation of the code. The tutorial can be found here:
https://unity3d.com/learn/tutorials/s/tanks-tutorial

This tutorial broke me into a lot of the basics of unity and exposed me
to a lot of fascinating possibilities. I learned the physics behind
trajectory based projectiles and how to apply gravity on objects in
Unity’s physics engine.

The assets used were provided by Unity itself and included in the
tutorial.

The game itself involves two players pitched against each other using
the same screen and keyboard. The players must try and destroy each
other. When a tank is destroyed the round ends. Each round won is worth
one point. Score is tracked and the first to five points wins.

— Controls —
Player 1:
– Movement: WASD
– Attack: Spacebar
Player 2:
– Movement: Arrow Keys
– Attack: Enter